no need to store production office in redux store
