--- public/video-ui/src/actions/WorkflowActions/getProductionOffice.js
@@ -1,6 +1,0 @@
-export function getProductionOffice() {
-  return {
-    type: 'WORKFLOW_GET_PRODUCTION_OFFICE',
-    receivedAt: Date.now()
-  };
-}
